Nettoyage exercices.js et exercices_5e.js
