+++ exercices/5e/5G51.js
@@ -1,0 +1,11 @@
+import Representer_un_solide from '../6e/6G41.js';
+/**
+ * référence  5G51
+ */
+export default function Representer_un_solide_5e() {
+	this.titre="Représenter un solide";
+	this.sup = 4;
+	this.classe = 5;
+	Representer_un_solide.call(this);
+  }
+
