+++ exercices/5e/5L13-5.js
@@ -1,0 +1,13 @@
+import Calculer_la_valeur_d_une_expression_litterale from './5L14.js'
+/**
+ * Caculer la valeur d'une expression littérale de degré 1 à une inconnue
+ * 5L13-5
+ * @author Sébastien Lozano forking 5L13 of Rémi Angot
+ */
+function Calculer_la_valeur_d_une_expression_litterale_deg1_inc1() {
+	Calculer_la_valeur_d_une_expression_litterale.call(this);
+	this.version = "5L13-5";
+	this.titre = "Calculer la valeur d'une expression littérale de degré 1 à 1 inconnue";
+	this.nb_questions = 2;
+
+}
