+++ include/mathalea_exercices.js
@@ -3,25 +3,11 @@ var liste_des_exercices_disponibles = {
   //"6C99" : separation6C,
   "beta6G24": Test_main_levee,
   "5A11": Tableau_criteres_de_divisibilite,
-  "5C11": Traduire_une_phrase_par_une_expression,
-  "5C12": Calculer_une_expression_numerique,
-  "5G11": Transformations_5e,
-  "5G21-1": Constructibilite_des_triangles_longueurs,
   "5G20-1": Vocabulaire_des_triangles_5e,
-  "5G22": DroiteRemarquableDuTriangle,
   "5G30-1": Egalite_d_angles,
-  "5G31-1": Constructibilite_des_triangles_angles,
-  "5G51": Representer_un_solide_5e,
-  "5N11-1": Tableaux_et_pourcentages_prix_constant,
-  "5N11-2": Tableaux_et_pourcentages_pourcent_constant,
-  "5N13": Exercice_fractions_simplifier,
   "5L14-1": Calculer_une_expression_litterale,
   "5L14-2": Exercice_substituer,
   "5L14-3": Traduire_une_phrase_par_une_expression_litterale_et_calculer,
-  "5L14-5": Calculer_la_valeur_d_une_expression_litterale_deg1_inc1,
-  "5L14-6": Determiner_derniere_operation_exp_num,
-  "5P10": Tableaux_et_proportionnalite,
-  "5R10-0": Trouver_oppose,
   "5R12": Reperage_point_du_quart_de_plan,
   "beta4C30": Comparer_puissance10,
   "beta3L11-5": identites_calculs,
